test(host): split GitHub adapter command contracts
